feat: 分页行数持久化(B22)全项目34个列表页修复 + 开发规范V8.37自洽性修正
- B21: 原锁计划及执行页面 MudSwitch fire-and-forget 改为 async EventCallback
- B22: 全项目34个ServerData列表页添加 _pageSize = state.PageSize 修复排序/筛选后行数复位
- ProductionRecords/ChemicalCompositions: 硬编码 RowsPerPage="10" 改为绑定 _pageSize
- 04_开发规范.md V8.37: 修复7处自洽性问题(模板代码/附录/AI检查项)
- SKILL.md 同步至 V8.37
- ConfigParameter/SectionFlowAnalysis/WorkOrderSchedule 新实体+页面+服务
- StandardWorkDay/StandardWorkDayDeliveryState 配置页内联编辑实现

+++ MES.Services/Mapping/DtoMapper.cs
@@ -80,7 +80,7 @@ public static class DtoMapper
         TotalWeight = entity.TotalWeight,
         DeliveryState = entity.DeliveryState,
         TotalItemCount = entity.TotalItemCount,
-        Status = (int)entity.Status,
+        Status = entity.Status,
         MaterialPlanStatus = (int)entity.MaterialPlanStatus,
         MaterialPlanRate = entity.MaterialPlanRate,
         CreatedTime = entity.CreatedTime
@@ -94,7 +94,7 @@ public static class DtoMapper
         ProductionMainNo = entity.ProductionMainNo,
         ProductionSubNo = entity.ProductionSubNo,
         OrderItemIds = entity.OrderItemIds,
-        Status = (int)entity.Status,
+        Status = entity.Status,
         SignDate = entity.SignDate,
         Salesman = entity.Salesman,
         EndCustomer = entity.EndCustomer,
